services/analyticsService/requirements.txt
alembic==1.16.4
fastapi==0.116.1
httpx==0.28.1
PyJWT==2.10.1
pandas==2.3.3
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.1.1
SQLAlchemy==2.0.43
uvicorn[standard]==0.35.0
